Move code in Array supporting find to a separate class
Main goal is to make Array slimmer and reduce the dependency on all
the templates supporting quey on integers.
