+++ Package.swift
@@ -46,6 +46,7 @@ let notSyncServerSources: [String] = [
     "realm/alloc.cpp",
     "realm/alloc_slab.cpp",
     "realm/array.cpp",
+    "realm/array_with_find.cpp",
     "realm/array_backlink.cpp",
     "realm/array_binary.cpp",
     "realm/array_blob.cpp",
@@ -351,6 +352,7 @@ let headers: [String] = [
     "realm/alloc.hpp",
     "realm/alloc_slab.hpp",
     "realm/array.hpp",
+    "realm/array_with_find.hpp",
     "realm/array_backlink.hpp",
     "realm/array_basic.hpp",
     "realm/array_basic_tpl.hpp",
@@ -363,6 +365,7 @@ let headers: [String] = [
     "realm/array_direct.hpp",
     "realm/array_fixed_bytes.hpp",
     "realm/array_integer.hpp",
+    "realm/array_integer_tpl.hpp",
     "realm/array_key.hpp",
     "realm/array_list.hpp",
     "realm/array_mixed.hpp",
@@ -508,6 +511,7 @@ let headers: [String] = [
     "realm/parser/keypath_mapping.hpp",
     "realm/parser/query_parser.hpp",
     "realm/query.hpp",
+    "realm/query_state.hpp",
     "realm/query_conditions.hpp",
     "realm/query_conditions_tpl.hpp",
     "realm/query_engine.hpp",
