+++ src/realm/array_timestamp.hpp
@@ -21,6 +21,7 @@
 
 #include <realm/array_integer.hpp>
 #include <realm/timestamp.hpp>
+#include <realm/query_conditions.hpp>
 
 namespace realm {
