+++ src/realm/query_engine.hpp
@@ -94,6 +94,7 @@ AggregateState      State of the aggregate - contains a state variable that stor
 #include <realm/array_key.hpp>
 #include <realm/array_string.hpp>
 #include <realm/array_binary.hpp>
+#include <realm/array_integer_tpl.hpp>
 #include <realm/array_timestamp.hpp>
 #include <realm/array_decimal128.hpp>
 #include <realm/array_fixed_bytes.hpp>
@@ -410,14 +411,14 @@ protected:
     size_t find_all_local(size_t start, size_t end)
     {
         if (run_single()) {
-            m_leaf_ptr->template find<TConditionFunction>(m_value, start, end, 0, m_state, nullptr);
+            m_leaf_ptr->template find<TConditionFunction>(m_value, start, end, m_state, nullptr);
         }
         else {
             auto callback = [this](size_t index) {
                 auto val = m_source_column->get_any(index);
                 return m_state->match(index, val);
             };
-            m_leaf_ptr->template find<TConditionFunction>(m_value, start, end, 0, m_state, callback);
+            m_leaf_ptr->template find<TConditionFunction>(m_value, start, end, m_state, callback);
         }
 
         return end;
