+++ src/realm/table.hpp
@@ -769,9 +769,6 @@ private:
 
     uint64_t get_sync_file_id() const noexcept;
 
-    static size_t get_size_from_ref(ref_type top_ref, Allocator&) noexcept;
-    static size_t get_size_from_ref(ref_type spec_ref, ref_type columns_ref, Allocator&) noexcept;
-
     /// Create an empty table with independent spec and return just
     /// the reference to the underlying memory.
     static ref_type create_empty_table(Allocator&, TableKey = TableKey());
@@ -1275,14 +1272,6 @@ inline bool Table::operator==(const Table& t) const
 inline bool Table::operator!=(const Table& t) const
 {
     return !(*this == t); // Throws
-}
-
-inline size_t Table::get_size_from_ref(ref_type top_ref, Allocator& alloc) noexcept
-{
-    const char* top_header = alloc.translate(top_ref);
-    std::pair<int_least64_t, int_least64_t> p = Array::get_two(top_header, 0);
-    ref_type spec_ref = to_ref(p.first), columns_ref = to_ref(p.second);
-    return get_size_from_ref(spec_ref, columns_ref, alloc);
 }
 
 inline bool Table::is_link_type(ColumnType col_type) noexcept
